+++ app/components/layout/layout-main.js
@@ -16,6 +16,7 @@ import Fab from '../shared/fab';
 import { common } from '../../styles/styles';
 import routerMain from '../routes/router-main';
 import routerModal from '../routes/router-modal';
+import { clientApp } from '../../lib/icebear';
 
 const { height } = Dimensions.get('window');
 
@@ -64,7 +65,6 @@ export default class LayoutMain extends SafeComponent {
         const pages = routerMain.pages;
         const currentPage = pages[routerMain.currentIndex];
         const currentComponent = routerMain.currentComponent;
-        console.log(currentComponent);
         const { actionsBar, showInput, leftIcon, rightIcon, layoutTitle } = currentComponent;
 
         const animatedBlock = (
@@ -89,7 +89,7 @@ export default class LayoutMain extends SafeComponent {
                 testID="mainLayout"
                 style={[common.container.root]}>
                 {animatedBlock}
-                <ProgressOverlay enabled={routerMain.loading} />
+                <ProgressOverlay enabled={routerMain.loading || clientApp.updatingAfterReconnect} />
                 <StatusBar
                     barStyle={uiState.externalViewer || routerModal.isBlackStatusBar ? 'default' : 'light-content'}
                     hidden={false} />