+++ app/components/layout/layout1.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { observer } from 'mobx-react/native';
 import { View, ScrollView } from 'react-native';
-import { observable } from 'mobx';
+import { observable, reaction, when } from 'mobx';
 import { MenuContext } from 'react-native-popup-menu';
 import SafeComponent from '../shared/safe-component';
 import SnackBarConnection from '../snackbars/snackbar-connection';
@@ -14,10 +14,33 @@ export default class Layout1 extends SafeComponent {
 
     componentDidMount() {
         uiState.currentScrollView = this._scrollView;
+        this.keyboardReaction = reaction(() => [uiState.keyboardHeight, uiState.focusedTextBox], () => {
+            if (!this.props.autoScroll) return;
+            // console.debug('layout1.js: keyboard height or focused textbox changed');
+            if (uiState.focusedTextBox) {
+                // console.debug('layout1.js: trying to measure textbox');
+                uiState.focusedTextBox.measure((fx, fy, width, height, px, py) => {
+                    const padding = height + uiState.height / 8;
+                    const preferrableY = uiState.height - padding - uiState.keyboardHeight;
+                    // console.debug(`layout1.js: preferrable Y: ${preferrableY}`);
+                    // console.debug(`layout1.js: py: ${py}`);
+                    if (uiState.keyboardHeight > 0) {
+                        if (py > preferrableY) {
+                            this._scrollView.scrollTo({ y: py - preferrableY });
+                            when(() => uiState.keyboardHeight === 0, () => this._scrollView.scrollTo({ y: 0 }));
+                        }
+                        if (py < 0) {
+                            this._scrollView.scrollTo({ y: 0 });
+                        }
+                    }
+                });
+            }
+        });
     }
 
     componentWillUnmount() {
         uiState.currentScrollView = null;
+        this.keyboardReaction();
     }
 
     renderThrow() {
@@ -75,6 +98,7 @@ Layout1.propTypes = {
     header: PropTypes.any,
     padding: PropTypes.any,
     noScroll: PropTypes.bool,
+    autoScroll: PropTypes.bool,
     noFitHeight: PropTypes.bool,
     noAutoHide: PropTypes.bool,
     noKeyboard: PropTypes.bool,